change ID types to long

+++ Bale.Net/Types/Update.cs
@@ -6,7 +6,7 @@ namespace Bale.Net.Types;
 public sealed class Update
 {
     [JsonPropertyName("update_id")]
-    public int UpdateId { get; set; }
+    public long UpdateId { get; set; }
 
     [JsonPropertyName("message")]
     public Message? Message { get; set; }
